Update patch set 2
Patch Set 2: Code-Review-1

Marcus, I don't think this is the way to go.  I think we should bump the version of so/libs to 1.2.1 and ask LF to release it.  This is now in progress with SO-536

Patch-set: 2
Reviewer: Gerrit User 1423 <1423@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1