Update patch set 2
Patch Set 2: -Code-Review

You raise a good point, Marcus. Leave this open and I'll do some more testing.  Because these are libraries used in the openstack adapter springboot app, we are rather limited to what springboot 1.5.13.RELEASE can support in terms of different jackson versions.  I do know that this version of springboot will not work with any jackson 2.9 version, but I think it will work with 2.8.10 or 2.8.11, which may help us somewhat with the CVEs.

Patch-set: 2